requirements.txt
Flask
azure-storage-blob
python-dotenv
Pillow